Fetch ProfileCard Data on page level
